refactor(desktop): move providers and workflow components into features/
providers: ProvidersChip, ProvidersPanel, PricingDialog, CostBadge,
TelemetryPill. workflow: WorkflowNextStepCta, NextActionChips,
StepOverrideRow. tests colocated.
