+++ apps/desktop/src/components/StatusBar.tsx
@@ -1,6 +1,6 @@
 import { CircleHelp } from 'lucide-react';
 import { useCurrentSession, useCurrentWorkspace } from '../store';
-import { TelemetryPill } from './TelemetryPill';
+import { TelemetryPill } from '../features/providers/components/TelemetryPill';
 
 interface StatusBarProps {
   onFocusWorkspaces?: () => void;
