+++ apps/desktop/src/features/providers/components/ProvidersPanel/index.tsx
@@ -2,11 +2,14 @@ import { useState } from 'react';
 import { Code2, Info, MousePointer2, RotateCw, Sparkles } from 'lucide-react';
 import type { LucideIcon } from 'lucide-react';
 import { Tooltip, cn } from '@kay-am/ui';
-import type { ProviderConnectionState, ProviderInfo } from '../features/providers/providers';
-import { providerAction } from '../features/providers/providers';
-import type { ProviderId } from '../features/providers/providers';
-import { useAppStore } from '../store';
-import { openUrl } from '../shared/lib/editor';
+import type {
+  ProviderConnectionState,
+  ProviderInfo,
+} from '../../../../features/providers/providers';
+import { providerAction } from '../../../../features/providers/providers';
+import type { ProviderId } from '../../../../features/providers/providers';
+import { useAppStore } from '../../../../store';
+import { openUrl } from '../../../../shared/lib/editor';
 
 const STATE_LABEL: Record<ProviderConnectionState, string> = {
   connected: 'Connected',
